[Authentication] Rename components to disambiguate from authorization.
Also made sure all stories are grouped together.

* Authentication is about identifying a person (sign in/up)
* Authorization is about whether a person has access to a resource.
* Auth is an ambiguous short-hand, the proper ones are authn and authz,
  but even better is no short-hand versions at all.
