+++ src/Components/Authentication/Mobile/__stories__/RegisterForm.story.tsx
@@ -1,0 +1,15 @@
+import { storiesOf } from "@storybook/react"
+import * as React from "react"
+
+import { MobileRegisterForm } from "../RegisterForm"
+
+storiesOf("Components/Authentication/Mobile", module).add(
+  "RegisterForm",
+  () => (
+    <MobileRegisterForm
+      values={{}}
+      handleSubmit={() => null}
+      handleTypeChange={() => mode => null}
+    />
+  )
+)
