--- src/Components/Authorization/Mobile/__stories__/RegisterForm.story.tsx
@@ -1,12 +1,0 @@
-import { storiesOf } from "@storybook/react"
-import * as React from "react"
-
-import { MobileRegisterForm } from "../RegisterForm"
-
-storiesOf("Components/Auth/Mobile", module).add("RegisterForm", () => (
-  <MobileRegisterForm
-    values={{}}
-    handleSubmit={() => null}
-    handleTypeChange={() => mode => null}
-  />
-))
